brought user account branch up to date with main. Still getting responses on user login and registration. Commented out UserServiceImp.java for the time being was causing program to fail during boot

+++ tenttalk-api/src/main/java/org/codestars/tenttalk_api/models/data/CampgroundRepository.java
@@ -6,8 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CampgroundRepository extends JpaRepository<Campground, Integer> {
-
-
-
+public interface CampgroundRepository extends JpaRepository<Campground, Long> {
 }
